Separate region inference logic from error handling better
